python/requirements.txt
requests==2.30.0
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pydantic==2.4.2
openai==0.28.1
python-dotenv==1.0.0
pytest-asyncio==0.21.1
pyright==1.1.335
pandas==2.1.2
chardet==5.2.0
scikit-learn==1.3.2
result==0.13.1
langchain==0.0.325
tiktoken==0.5.1
pinecone-client==2.2.4
python-aiconfig==1.0.4
